Reorganize code, removing test_invoice_factory.py and adding test_invoice_line_items.py
